server/requirements.txt
# Porygon Meal Planner API - Flask (Python)
flask>=3.0.0
flask-cors>=4.0.0
gunicorn>=21.0.0
flask-sqlalchemy>=3.1.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
firebase-admin>=6.2.0
google-cloud-aiplatform>=1.38.0
